Split tests into dedicated file, add comments
